chore: move unit tests
