+++ __tests__/unit/index.test-d.ts
@@ -1,6 +1,6 @@
 import { expectTypeOf, it } from "vitest"
 
-import * as lib from "../src"
+import * as lib from "../../src"
 
 it("returns void", () => {
 	expectTypeOf(lib.server).toBeObject()
