+++ __tests__/unit/index.test.ts
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest"
 
-import * as lib from "../src"
+import * as lib from "../../src"
 
 // TODO: Dummy test, meant to be removed when real tests come in
 it("exports something", () => {
