+++ vite.config.ts
@@ -18,7 +18,8 @@ export default defineConfig({
 			provider: "v8",
 			reporter: ["lcovonly", "text"],
 		},
-		setupFiles: ["./test/__setup__.ts"],
+		setupFiles: ["./__tests__/unit/__setup__.ts"],
+		include: ["./__tests__/unit/**/*.test.ts"],
 	},
 	define: {
 		__PACKAGE_VERSION__: JSON.stringify(pkg.version),
